# docs/openapi.yml
openapi: 3.1.0
info:
  title: Bank REST Test Assignment
  description: |-
    API specification for a bank cards web service.
    
    Features:
    - Bank cards CRUD with filters & pagination
    - Card operations (activating, blocking, transfers)
    - JWT authorization with access & refresh tokens
  version: 1.0.0
servers:
  - url: https://localhost:8443/api/v1
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Get users
      description: |-
        Get a list of users using filters and pagination.
        Available only to ROLE_ADMIN.
      operationId: getUsers
      security:
        - bearerAuth: [ROLE_ADMIN]
      parameters:
        - name: page
          in: query
          description: Number of the page. Default is 0.
          required: false
          schema:
            type: integer
            format: int32
        - name: size
          in: query
          description: Size of the page. Default is 10.
          required: false
          schema:
            type: integer
            format: int32
        - name: sort
          in: query
          description: |-
            Name of the field by which the sorting is going to be applied
            with an optional direction. Can be used multiple times to
            apply sorting by multiple fields. Default sorting is by id with
            ascending direction.
            Possible options: {id, firstName, lastName, email}
          required: false
          schema:
            type: string
            example: "id_asc"
        - name: firstName
          in: query
          description: Filter for the first name field (exact match)
          schema:
            type: string
            example: "ROBERT"
        - name: firstName_like
          in: query
          description: Filter for the first name field (by substring)
          schema:
            type: string
            example: "ROB"
        - name: lastName
          in: query
          description: Filter for the last name field (exact match)
          schema:
            type: string
            example: "MARTIN"
        - name: lastName_like
          in: query
          description: Filter for the last name field (by substring)
          schema:
            type: string
            example: "TIN"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'
        401:
          description: User unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        403:
          description: Insufficient privileges
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /users/{id}:
    delete:
      tags:
        - Users
      summary: Delete user
      description: Delete a userd by id. Available only to ROLE_ADMIN.
      operationId: deleteUser
      security:
        - bearerAuth: [ROLE_ADMIN]
      parameters:
        - name: id
          in: path
          schema:
            type: integer
            format: int64
            example: 1
          required: true
      responses:
        204:
          description: Successful operation
        401:
          description: User unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        403:
          description: Insufficient privileges
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
    put:
      tags:
        - Users
      summary: Update user
      description: Update a userd by id. Available only to ROLE_ADMIN.
      operationId: updateUser
      security:
        - bearerAuth: [ROLE_ADMIN]
      parameters:
        - name: id
          in: path
          schema:
            type: integer
            format: int64
            example: 1
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        required: true
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        400:
          description: Bad request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        401:
          description: User unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        403:
          description: Insufficient privileges
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        409:
          description: Email already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
    get:
      tags:
        - Users
      summary: Get user
      description: Get a user by id. Available only to ROLE_ADMIN.
      operationId: getUser
      security:
        - bearerAuth: [ROLE_ADMIN]
      parameters:
        - name: id
          in: path
          schema:
            type: integer
            format: int64
            example: 1
          required: true
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        401:
          description: User unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        403:
          description: Insufficient privileges
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        404:
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AppError:
      type: object
      properties:
        code:
          type: integer
          format: int32
        messages:
          type: array
          items:
              type: string
        timestamp:
          type: string
          format: date-time
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
          example: "Robert"
        lastName:
          type: string
          example: "Martin"
        email:
          type: string
          example: "example@gmail.com"
        roles:
          type: array
          items:
            type: string
            enum:
              - ROLE_USER
              - ROLE_ADMIN
            example: "ROLE_USER"
...
    PageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
...
    UserRequest:
      type: object
      properties:
        firstName:
          type: string
          example: Robert
        lastName:
          type: string
          example: Martin
        email:
          type: string
          example: "example@gmail.com"
        roles:
          type: array
          items:
            type: string
            enum:
              - ROLE_USER
              - ROLE_ADMIN
      required:
        - firstName
        - lastName
        - email
        - roles